requirements.txt
aws-cdk-lib==2.87.0
constructs>=10.0.0,<11.0.0
requests>=2.27.1
boto3>=1.28.5
